backend/src/modules/users/services/ShowProfileService.ts
import User from "@modules/users/infra/typeorm/entities/User";
import AppError from "../../../shared/errors/AppErrors";
import IUsersRepository from "../repositories/IUsersRepository";
import { inject, injectable } from "tsyringe";

interface Request {
  user_id: string,
}

@injectable()
export default class ShowProfileService {

  constructor(
    @inject('UsersRepository')
    private usersRepository: IUsersRepository,
  ) { }

  public async execute({ user_id }: Request): Promise<User> {


    // confere se o usuário á existe
    const user = await this.usersRepository.findById(user_id);

    if (!user) {
      throw new AppError('User not found',)
    }

    return user;



  }

}
